validates-created

+++ UserRegisterSystem/FrmConfigUser.aspx.cs
@@ -13,15 +13,23 @@ namespace UserRegisterSystem
     public partial class FrmConfigUser : System.Web.UI.Page
     {
         private User userActive;
+        private List<TextBox> controlls = new List<TextBox>();
         protected void Page_Load(object sender, EventArgs e)
         {
-            if (Security.isErrorSessionActive(Session["Error"]))
-                Response.Redirect("FrmErrorPage.aspx", false);
-            if (!Security.activeSession(Session["userActive"]))
-                Response.Redirect("FrmErrorPage.aspx", false);
-            userActive = (User)Session["userActive"];
-            if(!IsPostBack)
-                chargeControlls();
+                if (Security.isErrorSessionActive(Session["Error"]))
+                    Response.Redirect("FrmErrorPage.aspx", false);
+                if (!Security.activeSession(Session["userActive"]))
+                    Response.Redirect("FrmErrorPage.aspx", false);
+                userActive = (User)Session["userActive"];
+                if (!IsPostBack)
+                    chargeControlls();
+                    controllsToList();
+        }
+        private void controllsToList()
+        {
+            controlls.Add(TxtEmail);
+            controlls.Add(TxtUserName);
+            controlls.Add(TxtPassword);
         }
         public void chargeControlls()
         {
@@ -40,22 +48,52 @@ namespace UserRegisterSystem
         }
         protected void BtnSave_Click(object sender, EventArgs e)
         {
-            UserData userData = new UserData();
-            userActive.emailProp= TxtEmail.Text;
-            userActive.userNameProp = TxtUserName.Text;
-            userActive.passwordProp = TxtPassword.Text;
-            userData.updateUser(userActive);
 
-            string Rute = MapPath("~/Pictures/Profile-" + userActive.idProp.ToString() + ".jpg");
-            if (File.Exists(@Rute))
+            try
             {
-                File.Delete(@Rute);
-                TxtImage.PostedFile.SaveAs(Rute);
-            }
-            else TxtImage.PostedFile.SaveAs(Rute);
+                UserData userData = new UserData();
+                if (Validating_TextBox())
+                {
+                    userActive.emailProp = TxtEmail.Text;
+                    userActive.userNameProp = TxtUserName.Text;
+                    userActive.passwordProp = TxtPassword.Text;
+                    userData.updateUser(userActive);
+                    string Rute = MapPath("~/Pictures/Profile-" + userActive.idProp.ToString() + ".jpg");
+                    if (File.Exists(@Rute))
+                    {
+                        File.Delete(@Rute);
+                        TxtImage.PostedFile.SaveAs(Rute);
+                    }
+                    else TxtImage.PostedFile.SaveAs(Rute);
 
-            chargeControlls();
+                    chargeControlls();
+                }
+                else { LblWarning.Visible = true; }
+            }
+            catch (Exception ex) { Session.Add("Error", ex.ToString()); }
         }
+
+        private bool Validating_TextBox()
+        {
+            bool Result = true;
+            foreach (TextBox txt in controlls)
+            {
+                if (!Result) return Result;
+                else if (txt.Text.Length == 0)
+                {
+                    Result = false;
+                    LblWarning.Text = "Todos los campos son requeridos";
+                }
+                else if (!Helper.validate_LongText(txt.Text, 5, 30))
+                {
+                    Result = false;
+                    LblWarning.Text = "Cantidad de caracteres incorrectos";
+                    txt.Text = "Incorrecto";
+                }
+            }
+            return Result;
+        }
+
 
         protected void BtnGoToDashBoard_Click(object sender, EventArgs e)
         {
